% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indclmplusIndividualData.R
\name{indclmplus.IndividualData}
\alias{indclmplus.IndividualData}
\title{Fit chain-ladder+ to reverse time triangles.}
\usage{
\method{indclmplus}{IndividualData}(
  IndividualData,
  hazard.model = "cox",
  tie = "efron",
  baseline = "spline"
)
}
\arguments{
\item{model:}{model to be chosen for the fitting.}

\item{tie:}{ties handling, default is efron approach.}

\item{baseline:}{handling the baseline hazard. Default is a spline.}
}
\value{
individual chain ladder plus fit.
}
\description{
This function fits and computes the reserves for the invidual clmplus models.
}
