% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndividualData.R
\name{IndividualData}
\alias{IndividualData}
\title{Individual data set}
\usage{
IndividualData(
  data,
  id = NULL,
  continuous_features,
  categorical_features,
  accident_period,
  calendar_period,
  input_time_granularity = 1/12,
  output_time_granularity = 1/4,
  years = 4,
  calendar_period_extrapolation = FALSE,
  continuous_features_spline = NULL,
  degrees_cf = 3,
  degrees_of_freedom_cf = 4,
  degrees_cp = 3,
  degrees_of_freedom_cp = 4
)
}
\arguments{
\item{data}{data.frame for the individual reserving. The number of development periods can be bigger than the number of accident periods}

\item{id}{column that contains the policy id.}

\item{continuous_features}{continuous features columns to be scaled.}

\item{categorical_features}{categorical features columns to be one-hot encoded.}

\item{accident_period}{string that contains the name of the column in data corresponding to accident_period}

\item{calendar_period}{string that contains the name of the column in data corresponding to the calendar_period.}

\item{input_time_granularity}{time unit of the input data with reference to a year}

\item{output_time_granularity}{time unit of the output data with reference to a year}

\item{calendar_period_extrapolation}{logical, whether a spline for calendar extrapolation should be considered in the cox model fit. Default is FALSE.}

\item{continuous_features_spline}{weather a spline for smoothing continuous features should be added.}

\item{degrees_of_freedom_cf}{degrees of freedom of the splines for smoothing continuous features.}

\item{degrees_of_freedom_cp}{degrees of freedom of the splines for smoothing the calendar period effect.}

\item{degree_cf}{degrees of the spline for smoothing continuous features.}

\item{degree_cp}{degrees of the spline for smoothing the calendar period effect.}
}
\value{
Pre-processed data ready for individual reserving.
}
\description{
This function inputs and pre-processes the data for the application.
}
\references{
Pittarello, G., Hiabu, M., & Villegas, A. M. (2023). Chain Ladder Plus: a versatile approach for claims reserving. arXiv preprint arXiv:2301.03858.
}
