% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperreticulate.R
\name{install_pyresurv}
\alias{install_pyresurv}
\title{Install Python Environment for ReSurv}
\usage{
install_pyresurv(
  ...,
  envname = "pyresurv",
  new_env = identical(envname, "pyresurv")
)
}
\arguments{
\item{...}{Additional arguments for `virtualenv_create`.}

\item{envname}{`character`. Name of the environment created. Default `pyresurv`.}

\item{new_env}{`logical`. If `TRUE`, any existing Python virtual environment and/or `conda` environment specified by `envname` is deleted first.}
}
\value{
No return value.
}
\description{
Install a Python environment that allows the user to apply the Neural Network (NN) models.
}
