% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oos_lkh.R
\name{oos_lkh}
\alias{oos_lkh}
\title{Compute the out-of-the sample likelihood}
\usage{
oos_lkh(object)
}
\arguments{
\item{object}{\code{ReSurvFit} object.}
}
\value{
Out of the sample likelihood
}
\description{
When the lower triangle data are available, this method computes the likelihood on the lower triangle.
}
