% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooslkhReSurvFit.R
\name{ooslkh.default}
\alias{ooslkh.default}
\title{Compute the out-of-the sample likelihood}
\usage{
\method{ooslkh}{default}(object, ...)
}
\arguments{
\item{object}{\code{ReSurvFit} object.}
}
\value{
Out of the sample likelihood
}
\description{
When the lower triangle data are available, this method computes the likelihood on the lower triangle.
}
