% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooslkhReSurvFit.R
\name{ooslkh.ReSurvFit}
\alias{ooslkh.ReSurvFit}
\title{Compute the out-of-the sample likelihood}
\usage{
\method{ooslkh}{ReSurvFit}(object, ...)
}
\arguments{
\item{object}{\code{ReSurvFit} object.}

\item{...}{Other arguments to pass to ooslkh.}
}
\value{
\code{numeric}, out-of-sample likelihood.
}
\description{
When the lower triangle data are available, this method computes the likelihood on the lower triangle.
}
