% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions.R
\docType{data}
\name{pkg.env}
\alias{pkg.env}
\title{Helper functions}
\format{
An object of class \code{environment} of length 63.
}
\usage{
pkg.env
}
\description{
This script contains the utils functions that are used in ReSurv.
}
\keyword{datasets}
