% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictindclmplus.R
\name{predict.indclmplus}
\alias{predict.indclmplus}
\title{Individual clmplus fit predictions}
\usage{
\method{predict}{indclmplus}(ReSurvFit, ...)
}
\arguments{
\item{indclmplus}{resurv model to use for forecasting}
}
\value{
reserve predictions.
}
\description{
Predict the reserve for an individual clmplus fit
}
