% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictReserve.R
\name{predictReserve.ReSurvFit}
\alias{predictReserve.ReSurvFit}
\title{Predict deterministic reserve table from a ReSurvFit object}
\usage{
\method{predictReserve}{ReSurvFit}(object, granularity = c("output", "input"), ...)
}
\arguments{
\item{object}{A ReSurvFit object.}

\item{granularity}{Character. Either "output" or "input".}

\item{...}{Additional arguments passed to predict.ReSurvFit().}
}
\value{
A data.table with columns AP, DP, CP, IBNR.
}
\description{
Predict deterministic reserve table from a ReSurvFit object
}
