% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_party.R
\name{print.ReSurvCV}
\alias{print.ReSurvCV}
\title{Print method for ReSurvCV objects}
\usage{
\method{print}{ReSurvCV}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ReSurvCV}.}

\item{...}{Additional arguments, currently ignored.}
}
\description{
Print method for ReSurvCV objects
}
