% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReSurv-package.R
\docType{package}
\name{ReSurv-package}
\alias{ReSurv-package}
\title{ReSurv: Individual claim-count reserving}
\description{
Fit reverse-time hazard models with \code{\link{ReSurv}}, following
preprocessing with \code{\link{IndividualDataPP}}. Use
\code{\link{ReSurvCV}} for hyperparameter selection,
\code{\link{predictReserve}} for count forecasts, and
\code{\link{Score_Reserving}} to evaluate realized claims.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edhofman/ReSurv}
  \item \url{https://edhofman.github.io/ReSurv/}
  \item Report bugs at \url{https://github.com/edhofman/ReSurv/issues}
}

}
\author{
\strong{Maintainer}: Emil Hofman \email{emil_hofman@hotmail.dk} [copyright holder]

Authors:
\itemize{
  \item Gabriele Pittarello \email{gabriele.pittarello@uniroma1.it} (\href{https://orcid.org/0000-0003-3360-5826}{ORCID}) [copyright holder]
  \item Munir Hiabu \email{mh@math.ku.dk} (\href{https://orcid.org/0000-0001-5846-667X}{ORCID}) [copyright holder]
}

}
\keyword{internal}
