% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indclmplusIndividualData.R
\name{ReSurv.IndividualData}
\alias{ReSurv.IndividualData}
\title{Fit chain-ladder+ to reverse time triangles.}
\usage{
\method{ReSurv}{IndividualData}(
  IndividualData,
  hazard.model = "cox",
  tie = "efron",
  baseline = "spline",
  continuous.features.scaling.method = "minmax",
  random_seed = 1,
  hparameters = list(),
  percentage_data_training = 0.8
)
}
\arguments{
\item{model:}{model to be chosen for the fitting.}

\item{tie:}{ties handling, default is efron approach.}

\item{baseline:}{handling the baseline hazard. Default is a spline.}
}
\value{
individual chain ladder plus fit.
}
\description{
This function fits and computes the reserves for the invidual clmplus models.
}
