% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReSurvCVIndividualData.R
\name{ReSurvCV.IndividualData}
\alias{ReSurvCV.IndividualData}
\title{K fold cross-validation of ReSurv model.}
\usage{
\method{ReSurvCV}{IndividualData}(
  IndividualData,
  model,
  hparameters_grid,
  folds,
  random_seed,
  print_every_n = NULL,
  nrounds = NULL,
  verbose.cv = F,
  verbose,
  early_stopping_rounds = NULL
)
}
\arguments{
\item{IndividualData}{IndividualData object to use for the ReSurv fit cross-validation.}

\item{hparameters_grid}{grid of the hyperparameters to cross-validate.}

\item{folds}{number of folds.}

\item{random_seed}{random seed to make the results replicable.}
}
\value{
Best ReSurv model fit.
}
\description{
This function computes a K fold cross-validation of a pre-specified ReSurv model for a given grid of parameters.
}
