% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperxgboost.R
\name{risks_in_the_tie}
\alias{risks_in_the_tie}
\title{Helper functions for xgboost}
\usage{
risks_in_the_tie(starts_i, stops, stops_i)
}
\description{
This script contains the utils functions that are used in ReSurv xgboost.
In particular, it finds for each contribution the individuals at risk
in the tie.
}
