% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryReSurvPredict.R
\name{summary.ReSurvPredict}
\alias{summary.ReSurvPredict}
\title{Summary of IBNR predictions}
\usage{
\method{summary}{ReSurvPredict}(object, granularity = "input")
}
\arguments{
\item{object}{"ReSurvPredict" object specifying hazard and development factors.}

\item{granularity}{\code{character}, specify if which granularity the summary should be on.
\itemize{
\item{\code{"input"}}
\item{\code{"output"}}
}
Default is \code{"input"}.}
}
\value{
summary of predictions
}
\description{
Gives overview of IBNR predictions
}
