% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{data_generator}
\alias{data_generator}
\title{Individual data generator}
\usage{
data_generator(
  ref_claim = 2e+05,
  time_unit = 1/12,
  years = 4,
  random_seed = 1964,
  yearly_exposure = 6000,
  yearly_frequency = 0.2
)
}
\arguments{
\item{ref_claim}{reference claim size}

\item{time_unit}{output time unit}

\item{years}{simulated number of years}

\item{yearly_exposure}{volume underwritten}

\item{yearly_frequency}{yearly frequency}
}
\value{
individual claim data
}
\description{
This function generates individual claim data.
}
\references{
Pittarello, G., Hiabu, M., & Villegas, A. M. (2023). Chain Ladder Plus: a versatile approach for claims reserving. arXiv preprint arXiv:2301.03858.
}
