% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{data_generator}
\alias{data_generator}
\title{Individual data generator}
\usage{
data_generator(
  ref_claim = 2e+05,
  time_unit = 1/12,
  years = 4,
  random_seed = 1964,
  yearly_exposure = 6000,
  yearly_frequency = 0.2
)
}
\arguments{
\item{ref_claim}{\code{integer}, reference claim size.}

\item{time_unit}{\code{numeric}, output time unit.}

\item{years}{\code{integer}, number of years to be simulated.}

\item{yearly_exposure}{\code{integer}, volume underwritten each year.}

\item{yearly_frequency}{\code{numeric}, yearly frequency.}
}
\value{
Individual claims data. It contains the following columns:
\itemize{
\item{\code{AT}: Accident month in continuous time.}
\item{\code{AM}: Accident month.}
\item{\code{RT}: Reporting month in continuous time.}
\item{\code{RM}: Reporting month.}
\item{\code{DT}: Development month in continuous time.}
\item{\code{DM}: Development month.}
\item{\code{DT_rev}:Development month in continuous reverse time.}
\item{\code{DM_rev}: Development month in reverse time. }
\item{\code{TR}: Truncation time. }
\item{\code{I}: Event indicator.  }
}
}
\description{
This function generates the monthly individual claims data in the accompanying methodological paper using the \code{SynthETIC} package.
This simple function allows to simulate from a sand-box to test out the \code{ReSurv} approach.
Some parameters of the simulation can be changed.
}
\examples{
## Not run
input_data <- data_generator(random_seed = 1964)



}
\references{
Pittarello, G., Hiabu, M., & Villegas, A. M. (2023). Chain Ladder Plus: a versatile approach for claims reserving. arXiv preprint arXiv:2301.03858.
}
