% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndividualData.R
\name{IndividualData}
\alias{IndividualData}
\title{Individual data set}
\usage{
IndividualData(
  data,
  id = NULL,
  continuous_features,
  categorical_features,
  accident_period,
  calendar_period,
  input_time_unit = 1/12,
  output_time_unit = 1/4,
  continuous_features_spline = TRUE,
  degrees_of_freedom = 4
)

IndividualData(
  data,
  id = NULL,
  continuous_features,
  categorical_features,
  accident_period,
  calendar_period,
  input_time_unit = 1/12,
  output_time_unit = 1/4,
  continuous_features_spline = TRUE,
  degrees_of_freedom = 4
)
}
\arguments{
\item{data:}{data.frame for the individual reserving. The number of development periods can be bigger than the number of accident periods}

\item{id:}{column that contains the policy id.}

\item{continuous_features:}{continuous features columns to be scaled.}

\item{categorical_features:}{categorical features columns to be one-hot encoded.}

\item{accident_period:}{string that contains the name of the column in data corresponding to accident_period}

\item{calendar_period:}{string that contains the name of the column in data corresponding to the calendar_period.}

\item{input_time_unit:}{time unit of the input data with reference to a year}

\item{output_time_unit:}{time unit of the output data with reference to a year}

\item{continuous_features_spline:}{weather a spline for smoothing continuous features should be added.}

\item{degrees_of_freedom:}{degrees of freedom of the splines for smoothing continuous features.}
}
\value{
Pre-processed data ready for individual reserving.

Pre-processed data ready for individual reserving.
}
\description{
This function inputs and pre-processes the data for the application.

This function inputs and pre-processes the data for the application.
}
\references{
Pittarello, G., Hiabu, M., & Villegas, A. M. (2023). Chain Ladder Plus: a versatile approach for claims reserving. arXiv preprint arXiv:2301.03858.

Pittarello, G., Hiabu, M., & Villegas, A. M. (2023). Chain Ladder Plus: a versatile approach for claims reserving. arXiv preprint arXiv:2301.03858.
}
