% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReSurvIndividualData.R, R/plotReSurvFit.R
\name{plot.ReSurvFit}
\alias{plot.ReSurvFit}
\title{Draft for plot of \code{ReSurvFit} models for simulated data.}
\usage{
\method{plot}{ReSurvFit}(object)

\method{plot}{ReSurvFit}(object)
}
\arguments{
\item{object}{"ResurvFit" object specifying start time, end time and status.}
}
\value{
predictions
}
\description{
This function plots the results from the ReSurv fits.
}
