% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReSurvFit.R
\name{plot.ReSurvFit}
\alias{plot.ReSurvFit}
\title{Plot for machine learning models}
\usage{
\method{plot}{ReSurvFit}(object)
}
\arguments{
\item{object}{"ResurvFit" object specifying start time, end time and status.}
}
\value{
predictions
}
\description{
This function plots the results from the ReSurv fits.
}
