% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictReSurvFit.R
\name{predict.ReSurvFit}
\alias{predict.ReSurvFit}
\title{Predict IBNR frequency}
\usage{
\method{predict}{ReSurvFit}(
  object,
  newdata = NULL,
  lower_triangular_output = FALSE,
  minimal_output = FALSE,
  check_value = 1.85,
  ...
)
}
\arguments{
\item{object}{\code{ResurvFit} object specifying start time, end time and status.}

\item{newdata}{\code{IndividualDataPP} object that contains new data to predict.}

\item{lower_triangular_output}{\code{logical}, if set to \code{TRUE} we add the predicted lower triangle in input and output granularity to the \code{predict.ReSurvFit} output.}

\item{minimal_output}{\code{logical}, if set to \code{TRUE} return a reduced prediction object.}

\item{check_value}{\code{numeric}, check hazard value on initial granularity, if above threshold we increase granularity to try and adjust the development factor.}

\item{...}{Additional arguments to pass to the predict function.}
}
\value{
A \code{ReSurvPredict} object with fitted predictions, long triangle outputs, predicted counts, and optional lower-triangle outputs.
}
\description{
This function predicts the results from the ReSurv fits.
}
