% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictReSurvFit.R
\name{predict.ReSurvFit}
\alias{predict.ReSurvFit}
\title{Predict IBNR frequency}
\usage{
\method{predict}{ReSurvFit}(object)
}
\arguments{
\item{object}{"ResurvFit" object specifying start time, end time and status.}
}
\value{
predictions
}
\description{
This function predicts the results from the ReSurv fits.
}
