% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictReSurvFit.R
\name{predict.ReSurvFit}
\alias{predict.ReSurvFit}
\title{ReSurv fit predictions}
\usage{
\method{predict}{ReSurvFit}(object, IndividualData)
}
\arguments{
\item{object}{resurv model to use for forecasting}

\item{IndividualData}{individual data to preprocess for the fit.}
}
\value{
reserve predictions.
}
\description{
Predict the reserve for a ReSurv fit
}
