% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictReSurvFit.R
\name{predict.ReSurvFit}
\alias{predict.ReSurvFit}
\title{Predict IBNR frequency}
\usage{
\method{predict}{ReSurvFit}(
  object,
  newdata = NULL,
  grouping_method = "probability",
  check_value = 1.85,
  ...
)
}
\arguments{
\item{object}{\code{ResurvFit} object specifying start time, end time and status.}

\item{newdata}{\code{IndividualDataPP} object that contains new data to predict.}

\item{grouping_method}{\code{character}, use probability or exposure approach to group from input to output development factors. Choice between:
\itemize{
\item{\code{"exposure"}}
\item{\code{"probability"}}
}
Default is \code{"exposure"}.}

\item{check_value}{\code{numeric}, check hazard value on initial granularity, if above threshold we increase granularity to try and adjust the development factor.}

\item{...}{Additional arguments to pass to the predict function.}
}
\value{
Predictions for the \code{ReSurvFit} model.
}
\description{
This function predicts the results from the ReSurv fits.
}
