% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indclmplusIndividualData.R
\name{ReSurv.default}
\alias{ReSurv.default}
\title{Fit individual chain ladder plus models.}
\usage{
\method{ReSurv}{default}(
  IndividualData,
  hazard.model = "cox",
  tie = "efron",
  baseline = "spline",
  hparameters = list(),
  seed = 1
)
}
\arguments{
\item{model:}{model to be chosen for the fitting.}

\item{tie:}{ties handling. Default is efron approach.}

\item{baseline:}{handling the baseline hazard. Default is a spline.}
}
\value{
individual chain ladder plus fit.
}
\description{
This function fits and computes the reserves for the individual clmplus models.
}
