% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReSurvIndividualData.R
\name{ReSurv}
\alias{ReSurv}
\title{Fit individual chain ladder plus models.}
\usage{
ReSurv(
  IndividualData,
  hazard_model = "cox",
  tie = "efron",
  baseline = "spline",
  continuous_features_scaling_method = "minmax",
  random_seed = 1964,
  hparameters = list(),
  percentage_data_training = 0.8,
  grouping_method = "exposure"
)
}
\arguments{
\item{IndividualData}{IndividualData object to use for the ReSurv fit.}

\item{hazard_model}{hazard model supported from our package, must be provided as a string. The model can be chosen from:
\itemize{
\item{'cox': Standard Cox model for the hazard.}
\item{'deep_surv': Deep Survival Neural Network.}
\item{'xgboost': Gradient Boosting.}
}}

\item{tie}{ties handling, default is the Efron approach.}

\item{baseline}{handling the baseline hazard. Default is a spline.}

\item{continuous_features_scaling_method}{method to preprocess the features}

\item{random_seed}{random seed set for reproducibility}

\item{hparameters}{list of hyperparameters for 'deep_surv' and 'xgboost'. It will be disregarded for 'cox'.}

\item{percentage_data_training}{percentage of data used for training on the upper triangle.}

\item{grouping_method}{Use probability or exposure approach to group from input to output development factors.}
}
\value{
ReSurv fit.
}
\description{
This function fits and computes the reserves for the ReSurv models.
}
