% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResurvcvIndividualData.R
\name{ReSurvCV.IndividualData}
\alias{ReSurvCV.IndividualData}
\title{K fold cross-validation of ReSurv model.}
\usage{
\method{ReSurvCV}{IndividualData}(
  IndividualData,
  model,
  hparameters_grid,
  folds,
  random_seed,
  continuous_features_scaling_method = "minmax",
  print_every_n = 1L,
  nrounds = NULL,
  epochs = NULL,
  num_workers = 0,
  parallel = F,
  ncores = 1,
  verbose.cv = F,
  verbose = F,
  early_stopping_rounds = NULL
)
}
\arguments{
\item{IndividualData}{\code{IndividualData} object to use for the \code{ReSurv} fit cross-validation.}

\item{model}{\code{character}, machine learning for cross validation.}

\item{hparameters_grid}{\code{list}, grid of the hyperparameters to cross-validate.}

\item{folds}{\code{integer}, number of folds (i.e. K).}

\item{random_seed}{\code{integer}, random seed for making the code reproducible.}

\item{continuous_features_scaling_method}{\code{character}, method for scaling continuous features.}

\item{print_every_n}{\code{integer}, specific to the \code{xgboost} approach, see \code{xgboost::xgb.train} documentation.}

\item{nrounds}{\code{integer}, specific to \code{xgboost}, max number of boosting iterations.}

\item{epochs}{\code{integer}, specific to the \code{deepsurv} approach, epochs to be checked.}

\item{num_workers}{\code{numeric}, number of workers for the \code{deepsurv} approach, multi-process data loading with the specified number of loader worker processes.}

\item{parallel}{\code{logical}, specific to the \code{deepsurv} approach, whether to use parallel computing.}

\item{ncores}{\code{integer}, specific to \code{deepsurv}, max number of cores used.}

\item{verbose.cv}{\code{logical}, whether messages from cross-validation must be printed.}

\item{verbose}{\code{logical}, whether messages from the machine learning models must be printed.}

\item{early_stopping_rounds}{\code{integer}, specific to the \code{xgboost} approach, see \code{xgboost::xgb.train} documentation.}
}
\value{
Best ReSurv model fit.
}
\description{
This function computes a K fold cross-validation of a pre-specified ReSurv model for a given grid of parameters.
}
