% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResurvcvIndividualData.R
\name{ReSurvCV}
\alias{ReSurvCV}
\title{K fold cross-validation of ReSurv model.}
\usage{
ReSurvCV(
  IndividualData,
  model,
  hparameters_grid,
  folds,
  random_seed,
  continuous_features_scaling_method = "minmax",
  print_every_n = NULL,
  nrounds = NULL,
  early_stopping_rounds = NULL,
  epochs = 1,
  parallel = F,
  num_workers = 0,
  verbose.cv = F,
  verbose = F
)
}
\arguments{
\item{IndividualData}{IndividualData object to use for the ReSurv fit cross-validation.}

\item{hparameters_grid}{grid of the hyperparameters to cross-validate.}

\item{folds}{number of folds}
}
\value{
Best ReSurv model fit.
}
\description{
This function computes a K fold cross-validation of a pre-specified ReSurv model for a given grid of parameters.
}
