% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score_Reserving.R
\name{Score_Reserving}
\alias{Score_Reserving}
\title{Score reserving predictions}
\usage{
Score_Reserving(
  models,
  newdata,
  scoring_metrics = c("EI", "R-tot", "R-cell-wise", "R-cal-wise"),
  granularity = c("output", "input"),
  chain_ladder = TRUE,
  ...
)
}
\description{
Score reserving predictions
}
